HUD Scripts cleanup #2

+++ Assets/Scripts/HUD Scripts/SelectionBoxScript.cs
@@ -279,7 +279,7 @@ public class SelectionBoxScript : MonoBehaviour
             var lineRenderer = reticleRenderersByNode[current].GetComponent<LineRenderer>();
 
             reticleRenderersByNode[current].color = lineRenderer.startColor = lineRenderer.endColor
-                = Color.Lerp(new Color32((byte)100, (byte)100, (byte)100, (byte)255), Color.green,
+                = Color.Lerp(new Color32(100, 100, 100, 255), Color.green,
                     (1 - (delta.magnitude / originalDelta.magnitude)));
 
             if (delta.sqrMagnitude < PathAI.minDist)
